requirements.txt
streamlit
pandas
numpy
xgboost
shap
scikit-learn
google-generativeai